requirements.txt
aiohttp==0.21.0
appnope==0.1.0
blessings==1.6
chardet==2.3.0
click==5.1
colorama==0.3.3
cssselect==0.9.1
decorator==4.0.9
dj-database-url==0.4.0
Django==1.9.2
djangorestframework==3.3.2
falcon==0.3.0
gnureadline==6.3.3
humanize==0.5.1
ipdb==0.9.0
ipython==4.1.1
ipython-genutils==0.1.0
lxml==3.5.0
netifaces==0.10.4
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
pinax-cli==0.5
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.1.1
pyquery==1.2.11
python-dateutil==2.4.2
python-mimeparse==1.5.1
python-ptrace==0.8.1
requests==2.7.0
requests-cache==0.4.10
simplegeneric==0.8.1
six==1.10.0
traitlets==4.1.0
twtxt==1.0.1
wheel==0.26.0
